backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
requests==2.31.0
python-dotenv==1.0.1
playwright==1.42.0
httpx==0.27.0
pydantic==2.6.3 